CVAnalyser/backend/requirements.txt
ollama
flask
flask_cors
python-docx